+++ sampledb/scripts/__init__.py
@@ -14,6 +14,7 @@ from . import export_action_schema
 from . import create_other_user
 from . import send_announcement
 from . import set_administrator
+from . import set_user_readonly
 from . import run
 
 
@@ -29,5 +30,6 @@ script_functions = {
     'create_other_user': create_other_user.main,
     'send_announcement': send_announcement.main,
     'set_administrator': set_administrator.main,
+    'set_user_readonly': set_user_readonly.main,
     'run': run.main
 }