+++ sampledb/utils.py
@@ -37,6 +37,9 @@ def object_permissions_required(
                 return flask.abort(404)
             if not (logic.object_permissions.object_is_public(object_id) and required_object_permissions in Permissions.READ):
                 user_id = user_id_callable()
+                user = logic.users.get_user(user_id)
+                if user.is_readonly and required_object_permissions not in Permissions.READ:
+                    return on_unauthorized(object_id)
                 user_object_permissions = logic.object_permissions.get_user_object_permissions(object_id=object_id, user_id=user_id)
                 if required_object_permissions not in user_object_permissions:
                     return on_unauthorized(object_id)